ints to ntt and fft

+++ code/segtree.cpp
@@ -4,13 +4,13 @@ namespace seg {
     T t[2 * NN], id=0;
     T f(T a, T b) {return a+b;}
 
-    ll n=NN;  // array size
+    int n=NN;  // array size
 
-    void modify(ll p, T value) {  // set value at position p
+    void modify(int p, T value) {  // set value at position p
       for (p+=n, t[p] = value; p /= 2;) t[p] = f(t[2*p], t[2*p+1]);
     }
 
-    T query(ll l, ll r) { // fold f on interval [l, r)
+    T query(int l, int r) { // fold f on interval [l, r)
       T resl=id, resr=id;
       for (l += n, r += n; l < r; l /= 2, r /= 2) {
         if (l&1) resl = f(resl, t[l++]);
